+++ frontend/app/assets/js/en.Cfk8Eu-z.js
@@ -1,0 +1,1 @@
+import{n as e,t}from"./shared.1-CvUWxmFU.js";export{t as default,e as global};