+++ src/Shortcode/ScheduleSlider.php
@@ -4,6 +4,7 @@ namespace NabaHdwp\Shortcode;
 
 defined('ABSPATH') || exit;
 
+use NabaHdwp\Constant\PluginConstants;
 use NabaHdwp\Helper\TemplateEngine;
 use NabaHdwp\Model\Settings;
 use NabaHdwp\Service\VueService;
@@ -29,13 +30,14 @@ class ScheduleSlider
     {
         add_shortcode($name, function (array $atts): string {
             $data = [
-              'sessionData' => $this->settings->getSessionInitialData()
+              'sessionData' => $this->settings->getSessionInitialData(),
+              PluginConstants::INITIAL_DATA_KEY_GAMECENTER_BASE_URL => $this->settings->getGamecenterBaseUrl(),
             ];
 
             ob_start();
 
             TemplateEngine::render('/templates/shortcodes/schedule-slider.php', $data);
-            $this->vueService->enqueueAssets('compact');
+            $this->vueService->enqueueEmbedAssets();
 
             return ob_get_clean() ?: '';
         });